Extract Countdown to character package
