+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/core/domain/character/diseases/Disease.kt
@@ -1,20 +1,12 @@
 package cz.frantisekmasa.wfrp_master.common.core.domain.character.diseases
 
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
 import com.benasher44.uuid.uuid4
-import cz.frantisekmasa.wfrp_master.common.Plurals
-import cz.frantisekmasa.wfrp_master.common.Str
 import cz.frantisekmasa.wfrp_master.common.core.common.requireMaxLength
-import cz.frantisekmasa.wfrp_master.common.core.domain.NamedEnum
 import cz.frantisekmasa.wfrp_master.common.core.domain.character.CharacterItem
+import cz.frantisekmasa.wfrp_master.common.core.domain.character.Countdown
 import cz.frantisekmasa.wfrp_master.common.core.serialization.UuidAsString
-import dev.icerock.moko.parcelize.Parcelable
 import dev.icerock.moko.parcelize.Parcelize
-import dev.icerock.moko.resources.PluralsResource
-import dev.icerock.moko.resources.StringResource
-import dev.icerock.moko.resources.compose.stringResource
 import kotlinx.serialization.Serializable
 import cz.frantisekmasa.wfrp_master.common.compendium.domain.Disease as CompendiumDisease
 
@@ -75,29 +67,4 @@ data class Disease(
                 compendiumId = compendiumDisease.id,
             )
     }
-}
-
-@Parcelize
-@Serializable
-@Immutable
-data class Countdown(
-    val value: Int,
-    val unit: Unit,
-) : Parcelable {
-    init {
-        require(value >= 0) { "Remaining time must be non-negative" }
-    }
-
-    enum class Unit(
-        override val translatableName: StringResource,
-        val plural: PluralsResource,
-    ) : NamedEnum {
-        DAYS(Str.common_ui_units_days, Plurals.duration_days),
-        HOURS(Str.common_ui_units_hours, Plurals.duration_hours),
-        MINUTES(Str.common_ui_units_minutes, Plurals.duration_minutes),
-    }
-
-    @Composable
-    @Stable
-    fun toText(): String = stringResource(unit.plural, value, value)
 }
